Fix itemized schedule test
